requirements.txt
greenlet==2.0.2
numpy==1.25.0
pandas==2.0.3
psycopg2-binary==2.9.6
pyarrow==12.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.17
typing_extensions==4.7.1
tzdata==2023.3
